Massive Changes Check and Check mate flags
Ok so there are some massive fixes in this version. If anyone
experiances any problems at all with this new file let me know. I got
the handle for check, and checkmate in here. I have successfully run
like 30 runs to either stalemate or checkmate (95% of the time we are
getting stalemate). This required some fixes in everyones code so please
review the changes below and let me know if you have any problems with
them.

I suggest the next thing we work on is getting a slightly better
hueristic so that we can checkmate more often.

Also in playerInCheck(...) there are some commented out log lines. If
you have any trouble with the check or checkmate flags and you wish to
debug them on your own I suggest using these lines.